Move all type checking code to typecheck.hpp
