+++ src/ast.cpp
@@ -8,6 +8,14 @@ namespace {
 
 template<class... Ts> struct overloaded : Ts... { using Ts::operator()...; };
 
+auto verify(bool cond, std::string_view message)
+{
+    if (!cond) {
+        anzu::print(message);
+        std::exit(1);
+    }
+}
+
 template <typename T, typename Func>
 auto print_comma_separated(
     const std::vector<T>& values, Func&& formatter
@@ -36,11 +44,19 @@ auto print_node(const anzu::node_expr& root, int indent) -> void
             anzu::print("{}Variable: {}\n", spaces, node.name);
         },
         [&](const node_bin_op_expr& node) {
-            anzu::print("{}BinOp:\n", spaces);
+            anzu::print("{}BinOp\n", spaces);
             anzu::print("{}- Op: {}\n", spaces, node.op);
             anzu::print("{}- Lhs:\n", spaces);
+            if (!node.lhs) {
+                anzu::print("bin op has no lhs\n");
+                std::exit(1);
+            }
             print_node(*node.lhs, indent + 1);
             anzu::print("{}- Rhs:\n", spaces);
+            if (!node.rhs) {
+                anzu::print("bin op has no rhs\n");
+                std::exit(1);
+            }
             print_node(*node.rhs, indent + 1);
         },
         [&](const node_function_call_expr& node) {
