+++ src/parser.cpp
@@ -1,5 +1,6 @@
 #include "parser.hpp"
 #include "functions.hpp"
+#include "typecheck.hpp"
 #include "vocabulary.hpp"
 
 #include <optional>
@@ -14,8 +15,6 @@
 namespace anzu {
 namespace {
 
-template<class... Ts> struct overloaded : Ts... { using Ts::operator()...; };
-
 template <typename... Args>
 [[noreturn]] void parser_error(const token& tok, std::string_view msg, Args&&... args)
 {
@@ -23,114 +22,6 @@ template <typename... Args>
     anzu::print("[ERROR] ({}:{}) {}\n", tok.line, tok.col, formatted_msg);
     std::exit(1);
 }
-
-auto typeof(const anzu::object& object) -> std::string
-{
-    if (object.is<int>()) {
-        return std::string{tk_int};
-    }
-    if (object.is<bool>()) {
-        return std::string{tk_bool};
-    }
-    if (object.is<std::string>()) {
-        return std::string{tk_str};
-    }
-    if (object.is<object_list>()) {
-        return std::string{tk_list};
-    }
-    if (object.is<object_null>()) {
-        return std::string{tk_null_type};
-    }
-    return std::string{tk_any};
-}
-
-auto typeof_bin_op(
-    std::string_view lhs, std::string_view rhs, const token& op_token
-)
-    -> std::string
-{
-    const auto op = op_token.text;
-    const auto invalid_expr = [=]() {
-        parser_error(op_token, "could not evaluate '{} {} {}'", lhs, op, rhs);
-    };
-
-    if (lhs == tk_any || rhs == tk_any) {
-        return std::string{tk_any};
-    }
-
-    if (lhs != rhs) {
-        invalid_expr();
-    }
-
-    if (lhs == tk_list || lhs == tk_null_type) { // No support for having these in binary ops.
-        invalid_expr();
-    }
-
-    if (lhs == tk_str) {
-        if (op == tk_add) { // String concatenation
-            return std::string{tk_str};
-        }
-        invalid_expr();
-    }
-
-    if (lhs == tk_bool) {
-        if (op == tk_or || op == tk_and) {
-            return std::string{tk_bool};
-        }
-        invalid_expr();
-    }
-
-    if (is_comparison(op)) {
-        return std::string{tk_bool};
-    }
-    return std::string{tk_int};
-}
-
-struct parser_context
-{
-    anzu::tokenstream tokens;
-    std::unordered_map<std::string, function_signature> functions;
-    std::stack<std::unordered_map<std::string, std::string>> object_types;
-};
-
-auto fetch_function_signature(
-    const parser_context& ctx, const std::string& function_name
-)
-    -> function_signature
-{
-    if (auto it = ctx.functions.find(function_name); it != ctx.functions.end()) {
-        return it->second;
-    }
-
-    if (anzu::is_builtin(function_name)) {
-        return anzu::fetch_builtin(function_name).sig;
-    }
-
-    parser_error(ctx.tokens.curr(), "could not find function '{}'", function_name);
-}
-
-auto typeof_expr(const parser_context& ctx, const node_expr& expr) -> std::string
-{
-    return std::visit(overloaded {
-        [&](const node_literal_expr& node) {
-            return typeof(node.value);
-        },
-        [&](const node_variable_expr& node) {
-            return ctx.object_types.top().at(node.name);
-        },
-        [&](const node_function_call_expr& node) {
-            const auto& func_def = fetch_function_signature(ctx, node.function_name);
-            return func_def.return_type;
-        },
-        [&](const node_bin_op_expr& node) {
-            return typeof_bin_op(
-                typeof_expr(ctx, *node.lhs),
-                typeof_expr(ctx, *node.rhs),
-                node.op_token
-            );
-        }
-    }, expr);
-};
 
 template <typename Func>
 auto comma_separated_list(
@@ -287,7 +178,7 @@ auto parse_compound_factor(parser_context& ctx, std::int64_t level) -> node_expr
 auto parse_expression(parser_context& ctx) -> node_expr_ptr
 {
     auto ret = parse_compound_factor(ctx, std::ssize(bin_ops_table) - 1i64);
-    typeof_expr(ctx, *ret); // Typecheck the expression
+    type_of_expr(ctx, *ret); // Typecheck the expression
     return ret;
 }
 
@@ -303,7 +194,7 @@ auto parse_assigment_stmt(parser_context& ctx) -> node_stmt_ptr
     auto& stmt = node->emplace<anzu::node_assignment_stmt>();
     stmt.name = name;
     stmt.expr = parse_expression(ctx);
-    ctx.object_types.top()[stmt.name] = typeof_expr(ctx, *stmt.expr);
+    ctx.object_types.top()[stmt.name] = type_of_expr(ctx, *stmt.expr);
     return node;
 }
 
@@ -453,7 +344,7 @@ auto parse_function_call(parser_context& ctx) -> std::unique_ptr<NodeVariant>
     check_argc(token, out.function_name, std::ssize(sig.args), std::ssize(out.args));
     for (std::size_t idx = 0; idx != sig.args.size(); ++idx) {
         const auto& expected = sig.args.at(idx).type;
-        const auto& actual = typeof_expr(ctx, *out.args.at(idx));
+        const auto& actual = type_of_expr(ctx, *out.args.at(idx));
         if (expected != tk_any && expected != actual) {
             parser_error(
                 token, "invalid function call, arg {} expects type {}, got {}\n", idx, expected, actual
